# crates/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api",
    "description": "",
    "version": "0.1.0",
    "license": {
      "name": ""
    }
  },
  "paths": {
...
    "/projects/{project_id}/apps/": {
      "get": {
        "tags": ["app"],
        "operationId": "list_apps",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "replicas",
                      "image",
                      "port",
                      "deleted",
                      "projectId"
                    ],
                    "properties": {
                      "deleted": {
                        "type": "boolean"
                      },
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "image": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "port": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "projectId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "publicDomain": {
                        "type": "string"
                      },
                      "replicas": {
                        "type": "integer",
                        "format": "int32"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["app"],
        "operationId": "create_app",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppServiceSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "replicas",
                    "image",
                    "port",
                    "deleted",
                    "projectId"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "image": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "port": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "projectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "publicDomain": {
                      "type": "string"
                    },
                    "replicas": {
                      "type": "integer",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/apps/{app_id}/": {
      "get": {
        "tags": ["app"],
        "operationId": "get_app",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "replicas",
                    "image",
                    "port",
                    "deleted",
                    "projectId"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "image": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "port": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "projectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "publicDomain": {
                      "type": "string"
                    },
                    "replicas": {
                      "type": "integer",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["app"],
        "operationId": "delete_app",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "replicas",
                    "image",
                    "port",
                    "deleted",
                    "projectId"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "image": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "port": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "projectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "publicDomain": {
                      "type": "string"
                    },
                    "replicas": {
                      "type": "integer",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["app"],
        "operationId": "update_app",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialAppServiceSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "replicas",
                    "image",
                    "port",
                    "deleted",
                    "projectId"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "image": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "port": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "projectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "publicDomain": {
                      "type": "string"
                    },
                    "replicas": {
                      "type": "integer",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/apps/{app_id}/recover/": {
      "post": {
        "tags": ["app"],
        "operationId": "recover_app",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "replicas",
                    "image",
                    "port",
                    "deleted",
                    "projectId"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "image": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "port": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "projectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "publicDomain": {
                      "type": "string"
                    },
                    "replicas": {
                      "type": "integer",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/apps/{app_id}/status/": {
      "get": {
        "tags": ["app"],
        "operationId": "get_app_status",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "required": ["available", "state"],
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "state": {
                      "$ref": "#/components/schemas/AppReleaseState"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["BadRequest"]
              }
            }
          },
          {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["Validation"]
              }
            }
          },
          {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["AlreadyExists"]
              }
            }
          },
          {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["NotFound"]
              }
            }
          },
          {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["Unauthorized"]
              }
            }
          },
          {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["InternalError"]
              }
            }
          }
        ]
      },
      "AppReleaseState": {
        "type": "string",
        "enum": ["Unknown", "Failed", "Progressing", "Released"]
      },
...
      "AppServiceSchema": {
        "type": "object",
        "required": ["name", "replicas", "image", "port", "publicDomain"],
        "properties": {
          "image": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 65535
          },
          "publicDomain": {
            "$ref": "#/components/schemas/DomainName"
          },
          "replicas": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "DomainName": {
        "type": "object",
        "properties": {
          "subdomain": {
            "type": "string",
            "pattern": "(^[a-zA-Z0-9]$)|(^[a-zA-Z0-9][a-zA-Z0-9-]{0,61}[a-zA-Z0-9]$)",
            "minLength": 1,
            "maxLength": 62
          }
        }
      },
...
      "ErrorMessage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiError"
          },
          {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        ]
      },
      "PartialAppServiceSchema": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 65535
          },
          "publicDomain": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DomainName"
              }
            ]
          },
          "replicas": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "description": "token for protected content access",
        "name": "x-api-key",
        "in": "header"
      }
    }
  }
}